% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AirBox.R
\name{read.AirBox}
\alias{read.AirBox}
\title{Downlaod AirBox data}
\usage{
read.AirBox(x = 0)
}
\arguments{
\item{x}{numeric in {0, 1}; the type of AirBoxData to be downloaded.
\describe{
  \item{0:}{hourly instant AirBox data}
  \item{1:}{historical 7-day AirBox data}
  }}
}
\value{
a list containing AirBox data and device information
}
\description{
Downlaod AirBox data from Data.Taipei
}
\details{
The airbox data are dowload from Data.Taipe from the following link \link{http://data.taipei/opendata/datalist/datasetMeta?oid=4ba06157-3854-4111-9383-3b8a188c962a}
}
\examples{

# Download hourly instant AirBox data
system.time(d <-read.AirBox(0))

# 7 day average of PM 2.5 at elementary school

library(doBy)
library(ggplot2)
avg <- summaryBy(s_d0~school, data=d, FUN=c(mean, sd))
avg <- na.omit(avg)
od <- order(avg$s_d0.mean, decreasing=TRUE)
avg$school <- factor(avg$school, levels=avg$school[od])

Sys.setlocale(category = "LC_ALL", locale = "cht")

# Air quality ranks for elementary school in Taipei
ggplot(avg, aes(x=school, y=s_d0.mean)) +
  geom_point(stat = "identity")+
  geom_errorbar(aes(x=school, ymin=s_d0.mean-s_d0.sd, ymax=s_d0.mean+s_d0.sd))+
  xlab("")+ylab("PM 2.5")+
  coord_flip()+
  theme_bw()

# Time series of top 5 most polluted elementary school
sub <- subset(d, school==levels(avg$school)[1] | school==levels(avg$school)[2] | school==levels(avg$school)[3]| school==levels(avg$school)[4]| school==levels(avg$school)[5])
ggplot(data=sub, aes(x=time, y=s_d0, colour=school))+
  geom_line()+
  ylab("PM 2.5")+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 90, hjust = 1))
}
\author{
Chih-Lin Wei
}

