% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.LASS.R
\name{read.LASS}
\alias{read.LASS}
\title{Downlaod LASS data}
\usage{
read.LASS(x = 0)
}
\arguments{
\item{x}{numeric in {0, 1, 2,...}; the time of LASS data to be downloaded.
\describe{
  \item{0:}{Newest LASS data}
  \item{1:}{LASS data from 1 day before}
  \item{2:}{LASS data from 2 days before}
  \item{3:}{LASS data from 3 days before}
  }}
}
\value{
a list containing LASS data
}
\description{
Downlaod LASS data from \link{ftp://gpssensor.ddns.net:2121/}
}
\details{
The LASS data are dowload from Data.Taipe from the following link \link{ftp://gpssensor.ddns.net:2121/}
}
\examples{
lass <- read.LASS(x=0)
}
\author{
Chih-Lin Wei
}

